% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/du.R, R/du.ggplot.R, R/du.glm.R
\name{disk_usage}
\alias{disk_usage}
\alias{disk_usage.default}
\alias{disk_usage.ggplot}
\alias{disk_usage.glm}
\title{Disk Usage of Specific Object}
\usage{
disk_usage(object, full_data)

\method{disk_usage}{default}(object, full_data)

\method{disk_usage}{ggplot}(object = ggplot2::last_plot(), full_data)

\method{disk_usage}{glm}(object, full_data)
}
\arguments{
\item{object}{Different types, including \code{glm}, \code{ggplot}.}

\item{full_data}{Full data that the data passed into \code{object}
was derived from}
}
\value{
A binary data matrix.
}
\description{
Disk Usage of Specific Object
}
\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
utils::data(anorexia, package = "MASS")

anorex.1 <- glm(Postwt ~ Prewt + Treat + offset(Prewt),
                family = gaussian, data = anorexia[ anorexia$Prewt > 70.25,])
disk_usage(anorex.1, full_data = anorexia)
}
}
