% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/du.R
\name{du}
\alias{du}
\title{General Data Usage}
\usage{
du(..., full_data)
}
\arguments{
\item{...}{a set of objects to obtain data usage.}

\item{full_data}{Full data that the data passed into \code{object}
was derived from}
}
\value{
A data matrix of percentages.
}
\description{
General Data Usage
}
\examples{
## an example with offsets from Venables & Ripley (2002, p.189)
if (requireNamespace("MASS", quietly = TRUE)) {
utils::data(anorexia, package = "MASS")

sub_anor = anorexia[anorexia$Prewt > 73,]

anorex.1 <- glm(Postwt ~ Prewt + Treat + offset(Prewt),
                family = gaussian, data = anorexia[anorexia$Prewt > 70.25,])
anorex.2 <- glm(Postwt ~ Prewt + offset(Prewt),
                family = gaussian, data = anorexia)
du(anorex.1, anorex.2, sub_anor, full_data = anorexia)
library(ggplot2)
library(dplyr)
g = anorexia \%>\%
   filter(Prewt > 75 & Postwt < 100) \%>\%
   ggplot(aes(x = Prewt, y = Postwt, colour = Treat))
g2 = anorexia \%>\%
   ggplot(aes(x = Prewt, y = Postwt)) +
   geom_point() +
   geom_line(aes(size = Treat))
du(g, g2,  full_data = anorexia)
du(g, g2, anorex.1, anorex.2, list(data = sub_anor), full_data = anorexia)
}
}
