% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/du.R, R/du.gam.R, R/du.ggplot.R, R/du.glm.R
\name{disk_usage}
\alias{disk_usage}
\alias{disk_usage.default}
\alias{disk_usage.gam}
\alias{disk_usage.ggplot}
\alias{disk_usage.glm}
\alias{disk_usage.lm}
\title{Disk Usage of Specific Object}
\usage{
disk_usage(object, full_data)

\method{disk_usage}{default}(object, full_data)

\method{disk_usage}{gam}(object, full_data)

\method{disk_usage}{ggplot}(object = ggplot2::last_plot(), full_data)

\method{disk_usage}{glm}(object, full_data)

\method{disk_usage}{lm}(object, full_data)
}
\arguments{
\item{object}{Different types, including \code{glm}, \code{ggplot}.}

\item{full_data}{Full data that the data passed into \code{object}
was derived from}
}
\value{
A binary data matrix.
}
\description{
Disk Usage of Specific Object
}
\examples{
library(mgcv)
dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
b <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)
disk_usage(b, dat)


dat <- gamSim(1, n = 25000, dist = "normal", scale = 20)
bs <- "cr"
k <- 12
b <- bam(y ~ s(x0, bs = bs) + s(x1, bs = bs) + s(x2, bs = bs, k = k) +
           s(x3, bs = bs),
         data = dat[sample(nrow(dat), size = 10000),])
disk_usage(b, dat)
if (requireNamespace("MASS", quietly = TRUE)) {
  utils::data(anorexia, package = "MASS")

  anorex.1 <- glm(Postwt ~ Prewt + Treat + offset(Prewt),
                  family = gaussian, data = anorexia[ anorexia$Prewt > 70.25,])
  disk_usage(anorex.1, full_data = anorexia)
  disk_usage(list(data = anorex.1$data), full_data = anorexia)
}
if (requireNamespace("MASS", quietly = TRUE)) {
  utils::data(anorexia, package = "MASS")

  anorex.1 <- lm(Postwt ~ Prewt + offset(Prewt), data = anorexia)
  disk_usage(anorex.1, full_data = anorexia)
  disk_usage(list(model = anorex.1$model), full_data = anorexia)

}
}
